pyproject.toml
[tool.poetry]
name = "statistical-learning-language-clustering"
version = "0.1.0"
description = ""
authors = ["Connor McCarthy <mccarthy.connor.james@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
requests = "^2.25.1"
sklearn = "^0.0"
pdf2image = "^1.14.0"
plotly = "^4.14.3"
Pillow = "^8.1.2"
numpy = "^1.20.1"
pandas = "^1.2.3"
spacy = "^3.0.3"
hdbscan = "^0.8.27"
ipython = "^7.21.0"
ipykernel = "^5.5.0"
pytesseract = "^0.3.7"
en-core-web-lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.1/en_core_web_lg-2.3.1.tar.gz"}
nbformat = "^5.1.2"
wordfreq = "^2.4.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
mypy = "^0.800"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
make-setup = 'dephell deps convert --from-format=poetry --from-path=pyproject.toml --to-format=setuppy --to-path=setup.py'

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "pip", path = "requirements.txt"}

[tool.pytest.ini_options]
minversion = "6.2"
addopts = "--cov-report term-missing --cov=."
testpaths = [
    "tests",
]
